+++ Entities/DTOs/Details/HarcDetayDto.cs
@@ -16,6 +16,6 @@ namespace Entities.DTOs
         public string Turu { get; set; }
         public DateTime TahakkukTarihi { get; set; }
         public DateTime OdemeTarihi { get; set; }
-        public double Tutar { get; set; }
+        public Decimal Tutar { get; set; }
     }
 }
